[Add] Add verify the given xml string by given schema file [Update] Update slf4j version to 2.0.9 [Update] Update snakeyaml version to 2.2 [Release] Release version 1.2.0.3

+++ src/main/java/org/nervousync/generator/uuid/impl/UUIDv3Generator.java
@@ -62,7 +62,7 @@ public final class UUIDv3Generator extends UUIDGenerator {
             randomBytes[6] &= 0x0F;     /* clear version        */
             randomBytes[6] |= 0x30;     /* set to version 3     */
             randomBytes[8] &= 0x3F;     /* clear variant        */
-            randomBytes[8] |= 0x80;     /* set to IETF variant  */
+            randomBytes[8] |= (byte) 0x80;     /* set to IETF variant  */
             return new UUID(super.highBits(randomBytes), super.lowBits(randomBytes)).toString();
         } catch (NoSuchAlgorithmException e) {
             return Globals.DEFAULT_VALUE_STRING;
