+++ src/main/java/org/nervousync/generator/uuid/impl/UUIDv4Generator.java
@@ -51,7 +51,7 @@ public final class UUIDv4Generator extends UUIDGenerator {
         randomBytes[6] &= 0x0F;     /* clear version        */
         randomBytes[6] |= 0x40;     /* set to version 4     */
         randomBytes[8] &= 0x3F;     /* clear variant        */
-        randomBytes[8] |= 0x80;     /* set to IETF variant  */
+        randomBytes[8] |= (byte) 0x80;     /* set to IETF variant  */
         return new UUID(super.highBits(randomBytes), super.lowBits(randomBytes)).toString();
     }
     /**
